almost finished, calc page needs design

+++ src/pages/favorites/favorites.ts
@@ -59,30 +59,48 @@ export class FavoritesPage {
   }
 
   deleteFavs(fav) {
-    this.nativeStorage.getItem(this.session.getUser()).then(user => {
-      let favorites = user.fav
-      const password = user.password
-      const name = fav.name
-      var index = favorites.indexOf(fav.id)
-      if (index > -1) {
-        favorites.splice(index, 1)
-      }
-      this.nativeStorage.remove('fav').then(data => {
-        this.nativeStorage.setItem(this.session.getUser(), {
-          password,
-          fav: favorites
-        }).then(data => {
-          let alert = this.alertCtrl.create({
-            title: 'Fav deleted',
-            subTitle: `You have deleted ${name} from your favorites!` ,
-            buttons: ['Dismiss']
-          });
-          alert.present();
-          alert.onDidDismiss(() => {
-            this.navCtrl.setRoot(HomePage)
-          })
-        })
-      })
-    })
+    const name = fav.name
+    const confirm = this.alertCtrl.create({
+      title: 'You sure?',
+      message: `Do you want to delete ${name} from your favorites?`,
+      buttons: [
+        {
+          text: 'Disagree',
+          handler: () => {
+            console.log('Disagree clicked');
+          }
+        },
+        {
+          text: 'Agree',
+          handler: () => {
+            this.nativeStorage.getItem(this.session.getUser()).then(user => {
+              let favorites = user.fav
+              const password = user.password
+              var index = favorites.indexOf(fav.id)
+              if (index > -1) {
+                favorites.splice(index, 1)
+              }
+              this.nativeStorage.remove('fav').then(data => {
+                this.nativeStorage.setItem(this.session.getUser(), {
+                  password,
+                  fav: favorites
+                }).then(data => {
+                  let alert = this.alertCtrl.create({
+                    title: 'Fav deleted',
+                    subTitle: `You have deleted ${name} from your favorites!` ,
+                    buttons: ['Dismiss']
+                  });
+                  alert.present();
+                  alert.onDidDismiss(() => {
+                    this.navCtrl.setRoot(HomePage)
+                  })
+                })
+              })
+            })
+          }
+        }
+      ]
+    });
+    confirm.present();
   }
 }
